created a general context instead of placing it all in the app
