+++ src/components/Navigation-Build/Navigation-Build.js
@@ -1,13 +1,13 @@
 import React, { Component } from 'react';
 import { Link } from 'react-router-dom';
 import SearchBarBuild from '../SearchBar-Build/SearchBar-Build';
-import UserContext from '../../contexts/UserContext';
+import GeneralContext from '../../contexts/GeneralContext';
 import TokenService from '../../services/token-service';
 import jwtDecode from 'jwt-decode';
 
 export default class Navigation extends Component {
 
-  static contextType = UserContext;
+  static contextType = GeneralContext;
 
   handleLogoutClick = () => {
     TokenService.clearAuthToken();
