+++ src/components/TeamList-Edit/TeamList-Edit.js
@@ -1,10 +1,10 @@
 import React, { Component, Fragment } from 'react';
-import UserContext from '../../contexts/UserContext';
+import GeneralContext from '../../contexts/GeneralContext';
 import TeamEdit from '../Team-Edit/Team-Edit';
 
 export default class TeamList extends Component {
 
-  static contextType = UserContext;
+  static contextType = GeneralContext;
 
 
   render() {
