+++ src/components/TeamsButtonList-Edit/TeamsButtonList-Edit.js
@@ -1,10 +1,10 @@
 import React, { Component , Fragment} from 'react';
-import UserContext from '../../contexts/UserContext';
+import GeneralContext from '../../contexts/GeneralContext';
 import TeamButton from '../TeamButton/TeamButton';
 
 export default class TeamsButtonList extends Component {
 
-  static contextType = UserContext;
+  static contextType = GeneralContext;
 
   renderExpanded() {
