+++ src/components/TeamsButtonList-Public/TeamsButtonList-Public.js
@@ -1,10 +1,10 @@
 import React, { Component , Fragment} from 'react';
-import UserContext from '../../contexts/UserContext';
+import GeneralContext from '../../contexts/GeneralContext';
 import TeamButton from '../TeamButton/TeamButton';
 
 export default class TeamsButtonList extends Component {
 
-  static contextType = UserContext;
+  static contextType = GeneralContext;
 
   render() { // ** WILL NEED TO GO BACK AND ADD THE PAGINATION FEATURE **
