+++ src/index.js
@@ -4,11 +4,14 @@ import './index.css';
 import App from '../src/components/App/App';
 import * as serviceWorker from './serviceWorker';
 import {BrowserRouter} from 'react-router-dom';
+import { GeneralProvider } from '../src/contexts/GeneralContext';
 
 ReactDOM.render(
   <React.StrictMode>
     <BrowserRouter>
-      <App />
+      <GeneralProvider>
+        <App />
+      </GeneralProvider>
     </BrowserRouter>
   </React.StrictMode>,
   document.getElementById('root')
